ыав

+++ ShipForm/ClassSteamer.cs
@@ -12,7 +12,7 @@ namespace ShipForm
         public Color DopColor { private set; get; }
         public bool drawPipe;
         public bool drawWindow;
-        public ClassSteamer(int maxSpeed, float weight, Color mainColor, Color dopColor,
+        public ClassSteamer(int maxSpeed, int weight, Color mainColor, Color dopColor,
 bool drawWindow, bool drawPipe) :
  base(maxSpeed, weight, mainColor, 150, 150)
         {
